Renames pixiGrid to pixi-grid. Moves html-grid to html directory.
