+++ src/app/features/shared/grid/grid.module.ts
@@ -1,9 +1,9 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { BeatComponent } from "./beat.component";
-import { HtmlGridComponent } from './html-grid.component';
-import { PixiGridComponent } from './pixi/pixiGrid.component';
+import { BeatComponent } from "./html/beat.component";
+import { HtmlGridComponent } from './html/html-grid.component';
+import { PixiGridComponent } from './pixi/pixi-grid.component';
 
 
 @NgModule({
