+++ src/app/features/shared/grid/html/beat.component.scss
