+++ src/app/features/shared/grid/pixi/pixi-grid.component.scss
